+++ libubuntudownloadmanager/download_manager.h
@@ -42,8 +42,8 @@ public slots:
     QDBusObjectPath createDownload(const QString &url, const QVariantMap &metadata, const QVariantMap &headers);
     QDBusObjectPath createDownloadWithHash(const QString &url, const QString &algorithm, const QString &hash,
         const QVariantMap &metadata, const QVariantMap &headers);
-    uint defaultThrottle();
-    void setDefaultThrottle(uint speed);
+    qlonglong defaultThrottle();
+    void setDefaultThrottle(qlonglong speed);
     QList<QDBusObjectPath> getAllDownloads();
     QList<QDBusObjectPath> getAllDownloadsWithMetadata(const QString& name, const QString& value);
